Make it an error to def() an operator multiple times
Summary: Previously, if an operator is defined multiple times, it will be merged anyways. This code changes the behavior to spitting error if an operator is defined more than once.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Make it an error to def() an operator multiple times

Summary: Previously, if an operator is defined multiple times, it will be merged anyways. This code changes the behavior to spitting error if an operator is defined more than once.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 076a0c6eb6b656ef8e30d9315ae5e5a2300fec62
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39459